Adicionada propulsão no joystick (+)
+ Corrigir espaços para tabs (QtCreator tinha convertido)
+ Corrigir portabilidade da HAL
+ Atualizar mensagens do joystick e da interface para autoboat_msgs

+++ telemetria/include/telemetria/main_window.hpp
@@ -62,8 +62,8 @@ public Q_SLOTS:
     void on_campo_caracol_textChanged(QString qstr);
     void on_campo_dir_textChanged(QString qstr);
     void on_campo_esq_textChanged(QString qstr);
-    void on_vel_base_valueChanged(int valor);
-    void on_vel_caracol_valueChanged(int valor);
+    void on_vel_base_valueChanged(float valor);
+    void on_vel_caracol_valueChanged(float valor);
 
     // Caixas de seleção de direção
     void base_dir();
